refactor: inline single-symbol modules
