+++ openspec/changes/archive/2026-08-13-inline-single-symbol-modules/.openspec.yaml
@@ -1,0 +1,3 @@
+schema: spec-driven
+created: 2026-08-13
+skip_specs: true
