+++ openspec/changes/archive/2026-08-13-inline-single-symbol-modules/tasks.md
@@ -1,0 +1,9 @@
+## 1. Acceptance Test
+
+- [x] 1.1 Add import-contract tests for the new symbol locations and observe them fail.
+
+## 2. Implementation
+
+- [x] 2.1 Move `RuntimeOptions`, update CLI imports, and delete `cli/runtime.py`.
+- [x] 2.2 Move provider defaults, update consumers/tests, and delete `providers/constants.py`.
+- [x] 2.3 Run CLI/provider tests, lint, and spec-anchor checks.
