+++ tests/providers/test_factory.py
@@ -221,8 +221,7 @@ class TestDefaultTimeout:
         """The adapter's last-resort timeout applies whenever a caller builds a
         provider outside the factory (or passes timeout=None through), so it must
         not silently reimpose a budget shorter than the factory would have."""
-        from lgtmaybe.providers.constants import CLOUD_TIMEOUT
-        from lgtmaybe.providers.factory import default_timeout_for
+        from lgtmaybe.providers.factory import CLOUD_TIMEOUT, default_timeout_for
 
         # One definition, shared: >= would let the two 600s drift apart silently.
         assert CLOUD_TIMEOUT == min(default_timeout_for(p) for p in Provider)
